DIO: fix unsafe casts in native code for 64 bits architectures * Recompiled all native libraries and Java classes * Added two makefiles for ARMv6 and AArch64 architectures

+++ target-platform/jdk.dio-parent/src/dev/src/share/classes/com/oracle/dio/utils/Constants.java
@@ -32,7 +32,7 @@ public final class Constants {
     public static final String FACTORY_PREFIX = "com.oracle.dio.";
     public static final String FACTORY = "Factory";
     public static final String CONFIG = "Config";
-    public static final int INVALID_HANDLE = -1;
+    public static final long INVALID_HANDLE = -1;
     public static final String PATH = "deviceName";
     public static final String IMPL = ".impl.";
     public static final String TYPE = "type";
